Extract ShareRootViewController for not doing view / life cycle stuff in UINavigationController
