+++ LineSDK/LineSDK/SharingUI/ShareViewController.swift
@@ -70,88 +70,24 @@ public class ShareViewController: UINavigationController {
         static var navigationBarTextColor:  UIColor { return  .white }
     }
 
-    typealias ColumnIndex = ColumnDataStore<ShareTarget>.ColumnIndex
-
     public var navigationBarTintColor = Design.navigationBarTintColor { didSet { updateNavigationStyles() } }
     public var navigationBarTextColor = Design.navigationBarTextColor { didSet { updateNavigationStyles() } }
     public var statusBarStyle = Design.preferredStatusBarStyle { didSet { updateNavigationStyles() } }
 
     // Root & Data
-    private var rootViewController: UIViewController! { didSet { print("Set") } }
-    private var store: ColumnDataStore<ShareTarget>!
-
-    // States
-    @objc dynamic private var allLoaded: Bool = false
-
-    // Observers
-    private var selectingObserver: NotificationToken!
-    private var deselectingObserver: NotificationToken!
-
-    private var loadedObserver: NSKeyValueObservation?
+    private let rootViewController = ShareRootViewController()
 
     public weak var shareDelegate: ShareViewControllerDelegate?
 
-    private lazy var selectedTargetView = SelectedTargetView()
-
-    private var indicatorContainer: UIView?
-
     // MARK: - Initializers
     public init() {
-        let store = ColumnDataStore<ShareTarget>(columnCount: MessageShareTargetType.allCases.count)
-        let controllers = MessageShareTargetType.allCases.map { index -> ShareTargetSelectingViewController in
-            let controller = ShareTargetSelectingViewController(store: store, columnIndex: index.rawValue)
-            // Force load view for pages to setup table view initial state.
-            _ = controller.view
-            return controller
-        }
-
-        let pages = zip(MessageShareTargetType.allCases, controllers).map {
-            index, controller -> PageViewController.Page in
-            return .init(viewController: controller, title: index.title)
-        }
-
-        let rootViewController = PageViewController(pages: pages)
-
-        super.init(rootViewController: rootViewController)
-
-        self.store = store
-        self.rootViewController = rootViewController
-        controllers.forEach { $0.delegate = self }
-
+        super.init(nibName: nil, bundle: nil)
+        self.viewControllers = [rootViewController]
         updateNavigationStyles()
-    }
-
-    deinit {
-        ImageManager.shared.purgeCache()
-    }
-
-    @objc
-    func foo() {
-        selectedTargetView.setMode((selectedTargetView.mode == .show) ? .hide : .show,
-                                   animated: true)
-    }
-
-    public override init(nibName nibNameOrNil: String?, bundle nibBundleOrNil: Bundle?) {
-        super.init(nibName: nibNameOrNil, bundle: nibBundleOrNil)
     }
 
     required init?(coder aDecoder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
-    }
-
-    // MARK: - Lift Cycle
-    public override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        setupSubviews()
-        setupLayouts()
-
-        // default
-        selectedTargetView.setMode(.hide, animated: false)
-
-        // Wait for child view controllers setup themselves.
-        loadGraphList()
-        setupObservers()
     }
 
     // MARK: - Setup & Style
@@ -176,36 +112,6 @@ public class ShareViewController: UINavigationController {
         navigationBar.titleTextAttributes = [.foregroundColor: navigationBarTextColor]
     }
 
-    private func setupObservers() {
-        selectingObserver = NotificationCenter.default.addObserver(
-            forName: .columnDataStoreDidSelect, object: store, queue: nil)
-        {
-            [unowned self] noti in
-            self.handleSelectingChange(noti)
-        }
-
-        deselectingObserver = NotificationCenter.default.addObserver(
-            forName: .columnDataStoreDidDeselect, object: store, queue: nil)
-        {
-            [unowned self] noti in
-            self.handleSelectingChange(noti)
-        }
-    }
-  
-    private func setupSubviews() {
-        view.addSubview(selectedTargetView)
-    }
-
-    private func setupLayouts() {
-        selectedTargetView.translatesAutoresizingMaskIntoConstraints = false
-        NSLayoutConstraint.activate([
-            selectedTargetView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            selectedTargetView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
-            selectedTargetView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
-            selectedTargetView.topAnchor.constraint(equalTo: safeBottomAnchor,
-                                                    constant: -SelectedTargetView.Design.height)
-            ])
-    }
 }
 
 // MARK: - Controller Actions
@@ -214,64 +120,6 @@ extension ShareViewController {
         dismiss(animated: true) {
             self.shareDelegate?.shareViewControllerDidCancelSharing(self)
         }
-    }
-
-    private func loadGraphList() {
-
-        let friendsRequest = GetFriendsRequest(sort: .relation, pageToken: nil)
-        let chainedFriendsRequest = ChainedPaginatedRequest(originalRequest: friendsRequest)
-        chainedFriendsRequest.onPageLoaded.delegate(on: self) { (self, response) in
-            self.store.append(data: response.friends, to: MessageShareTargetType.friends.rawValue)
-        }
-
-        let groupsRequest = GetGroupsRequest(pageToken: nil)
-        let chainedGroupsRequest = ChainedPaginatedRequest(originalRequest: groupsRequest)
-        chainedGroupsRequest.onPageLoaded.delegate(on: self) { (self, response) in
-            self.store.append(data: response.groups, to: MessageShareTargetType.groups.rawValue)
-        }
-
-        let sendingDispatchGroup = DispatchGroup()
-
-        sendingDispatchGroup.enter()
-        Session.shared.send(chainedFriendsRequest) { result in
-            sendingDispatchGroup.leave()
-            switch result {
-            case .success:
-                break
-            case .failure(let error):
-                self.shareDelegate?.shareViewController(self, didFailLoadingListType: .friends, withError: error)
-            }
-        }
-
-        sendingDispatchGroup.enter()
-        Session.shared.send(chainedGroupsRequest) { result in
-            sendingDispatchGroup.leave()
-            switch result {
-            case .success:
-                break
-            case .failure(let error):
-                self.shareDelegate?.shareViewController(self, didFailLoadingListType: .groups, withError: error)
-            }
-        }
-
-        sendingDispatchGroup.notify(queue: .main) {
-            self.allLoaded = true
-        }
-    }
-
-    private func handleSelectingChange(_ notification: Notification) {
-        let count = store.selected.count
-        if count == 0 {
-            rootViewController.navigationItem.rightBarButtonItem = nil
-        } else {
-            let title = Localization.string("common.action.send")
-            rootViewController.navigationItem.rightBarButtonItem = UIBarButtonItem(
-                title: "\(title) (\(count))", style: .plain, target: self, action: #selector(sendMessage))
-        }
-    }
-
-    @objc private func sendMessage() {
-        print("Send")
     }
 }
 
@@ -298,51 +146,10 @@ extension ShareViewController {
     }
 }
 
-// MARK: - Selecting view controller delegate
-extension ShareViewController: ShareTargetSelectingViewControllerDelegate {
-    func shouldSearchStart(_ viewController: ShareTargetSelectingViewController) -> Bool {
-        if allLoaded {
-            return true
-        }
-
-        addLoadingIndicator()
-        loadedObserver = observe(\.allLoaded, options: .new) { [weak self] controller, change in
-            guard let self = self else { return }
-            if let loaded = change.newValue, loaded {
-                self.removeLoadingIndicator()
-                self.loadedObserver = nil
-                viewController.continueSearch()
-            }
-        }
-
-        return false
-    }
-
-    private func addLoadingIndicator() {
-        let container = UIView(frame: .zero)
-        let indicator = UIActivityIndicatorView(style: .whiteLarge)
-        indicator.color = .gray
-
-        indicator.translatesAutoresizingMaskIntoConstraints = false
-        container.addSubview(indicator)
-        NSLayoutConstraint.activate([
-            indicator.centerXAnchor.constraint(equalTo: container.centerXAnchor),
-            indicator.centerYAnchor.constraint(equalTo: container.centerYAnchor),
-        ])
-
-        // Add the loading indicator container to root view controller (page), so user has a chance to
-        // close the sharing UI by tapping "Close" button in the navigation bar.
-        rootViewController.view.addChildSubview(container)
-
-        indicator.startAnimating()
-
-        indicatorContainer = container
-    }
-
-    private func removeLoadingIndicator() {
-        guard let container = indicatorContainer else {
-            return
-        }
-        container.removeFromSuperview()
+extension ShareViewController: ShareRootViewControllerDelegate {
+    func shareViewController(_ controller: ShareRootViewController,
+                             didFailLoadingListType shareType: MessageShareTargetType,
+                             withError error: LineSDKError) {
+        shareDelegate?.shareViewController(self, didFailLoadingListType: shareType, withError: error)
     }
 }
